+++ apps/framework-cli/Cargo.toml
@@ -12,6 +12,7 @@ openssl = { version = "0.10", features = ["vendored"] }
 clap = { version = "4.3.17", features = ["derive"] }
 tokio = { version = "1", features = ["full"] }
 console = "0.15.7"
+thiserror = "1.0.59"
 hyper = { version = "1", features = ["full"] }
 hyper-tls = "0.6.0"
 notify = { version = "6.1.1", default-features = false, features = [
