+++ apps/framework-cli/src/framework/prisma.rs
@@ -1,0 +1,1 @@
+pub mod parser;
