--- apps/framework-cli/src/framework/sdks/mapper.rs
@@ -1,1 +1,0 @@
-
