+++ apps/framework-cli/src/infrastructure/olap/clickhouse/errors.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, thiserror::Error)]
+#[error("failed interact with clickhouse")]
+#[non_exhaustive]
+pub enum ClickhouseError {
+    #[error("Clickhouse - Unsupported data type: {type_name}")]
+    UnsupportedDataTypeError { type_name: String },
+}
