+++ apps/framework-cli/src/infrastructure/olap/clickhouse/mapper.rs
@@ -1,10 +1,11 @@
-use crate::framework::schema::{ColumnType, FieldArity, Table, TableType};
+use crate::framework::schema::{ColumnType, Table, TableType};
 
 use crate::infrastructure::olap::clickhouse::model::{
     ClickHouseColumn, ClickHouseColumnType, ClickHouseFloat, ClickHouseInt, ClickHouseTable,
     ClickHouseTableType,
 };
 
+use super::errors::ClickhouseError;
 use super::model::sanitize_column_name;
 
 pub fn clickhouse_table_type_mapper(table_type: TableType) -> ClickHouseTableType {
@@ -14,48 +15,53 @@ pub fn clickhouse_table_type_mapper(table_type: TableType) -> ClickHouseTableTyp
     }
 }
 
-pub fn std_field_type_to_clickhouse_type_mapper(field_type: ColumnType) -> ClickHouseColumnType {
+pub fn std_field_type_to_clickhouse_type_mapper(
+    field_type: ColumnType,
+) -> Result<ClickHouseColumnType, ClickhouseError> {
     match field_type {
-        ColumnType::String => ClickHouseColumnType::String,
-        ColumnType::Boolean => ClickHouseColumnType::Boolean,
-        ColumnType::Int => ClickHouseColumnType::ClickhouseInt(ClickHouseInt::Int64),
-        ColumnType::Float => ClickHouseColumnType::ClickhouseFloat(ClickHouseFloat::Float64),
-        ColumnType::Decimal => ClickHouseColumnType::Decimal,
-        ColumnType::DateTime => ClickHouseColumnType::DateTime,
-        ColumnType::Enum(x) => ClickHouseColumnType::Enum(x),
-        ColumnType::Unsupported => ClickHouseColumnType::Unsupported,
-        _ => ClickHouseColumnType::Unsupported,
+        ColumnType::String => Ok(ClickHouseColumnType::String),
+        ColumnType::Boolean => Ok(ClickHouseColumnType::Boolean),
+        ColumnType::Int => Ok(ClickHouseColumnType::ClickhouseInt(ClickHouseInt::Int64)),
+        ColumnType::Float => Ok(ClickHouseColumnType::ClickhouseFloat(
+            ClickHouseFloat::Float64,
+        )),
+        ColumnType::Decimal => Ok(ClickHouseColumnType::Decimal),
+        ColumnType::DateTime => Ok(ClickHouseColumnType::DateTime),
+        ColumnType::Enum(x) => Ok(ClickHouseColumnType::Enum(x)),
+        ColumnType::Array(inner_std_type) => {
+            let inner_clickhouse_type = std_field_type_to_clickhouse_type_mapper(*inner_std_type)?;
+            Ok(ClickHouseColumnType::Array(Box::new(inner_clickhouse_type)))
+        }
+        ColumnType::BigInt => Err(ClickhouseError::UnsupportedDataTypeError {
+            type_name: "BigInt".to_string(),
+        }),
+        ColumnType::Json => Err(ClickhouseError::UnsupportedDataTypeError {
+            type_name: "Json".to_string(),
+        }),
+        ColumnType::Bytes => Err(ClickhouseError::UnsupportedDataTypeError {
+            type_name: "Bytes".to_string(),
+        }),
     }
 }
 
-pub fn arity_mapper(arity: schema_ast::ast::FieldArity) -> FieldArity {
-    match arity {
-        schema_ast::ast::FieldArity::Required => FieldArity::Required,
-        schema_ast::ast::FieldArity::Optional => FieldArity::Optional,
-        schema_ast::ast::FieldArity::List => FieldArity::List,
+pub fn std_table_to_clickhouse_table(table: Table) -> Result<ClickHouseTable, ClickhouseError> {
+    let mut columns = Vec::new();
+    for column in table.columns {
+        let clickhouse_column = ClickHouseColumn {
+            name: sanitize_column_name(column.name),
+            column_type: std_field_type_to_clickhouse_type_mapper(column.data_type)?,
+            required: column.required,
+            unique: column.unique,
+            primary_key: column.primary_key,
+            default: None, // TODO: Implement the default mapper
+        };
+        columns.push(clickhouse_column);
     }
-}
 
-pub fn std_table_to_clickhouse_table(table: Table) -> ClickHouseTable {
-    let columns = table
-        .columns
-        .into_iter()
-        .map(|column| {
-            ClickHouseColumn {
-                name: sanitize_column_name(column.name),
-                column_type: std_field_type_to_clickhouse_type_mapper(column.data_type),
-                arity: column.arity,
-                unique: column.unique,
-                primary_key: column.primary_key,
-                default: None, // TODO: Implement the default mapper
-            }
-        })
-        .collect::<Vec<ClickHouseColumn>>();
-
-    ClickHouseTable {
+    Ok(ClickHouseTable {
         db_name: table.db_name,
         name: table.name,
         columns,
         table_type: clickhouse_table_type_mapper(table.table_type),
-    }
+    })
 }
